fix cleanup

+++ launcher/core/cleanup.go
@@ -2,32 +2,63 @@ package core
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"github.com/ExchangeUnion/xud-docker/launcher/types"
 	"golang.org/x/sync/errgroup"
 	"os"
 	"os/exec"
+	"runtime"
+	"strings"
+	"time"
 )
 
 func (t *Launcher) stopService(ctx context.Context, name string) error {
-	s, err := t.GetService("boltz")
+	t.Logger.Debugf("Stopping %s", name)
+	s, err := t.GetService(name)
 	if err != nil {
 		return err
 	}
 	if err := s.Stop(ctx); err != nil {
+		if strings.Contains(err.Error(), "No such container") {
+			t.Logger.Debugf("Service %s stopped already", name)
+			return nil
+		}
 		return err
 	}
+	for {
+		status, err := s.GetStatus(ctx)
+		if err != nil {
+			return err
+		}
+		t.Logger.Debugf("%s: %s", name, status)
+
+		if status == "Container exited" {
+			break
+		}
+
+		select {
+		case <-ctx.Done():
+			return errors.New("interrupted")
+		case <-time.After(3 * time.Second):
+		}
+	}
+
 	return nil
 }
 
 func (t *Launcher) stopAll(ctx context.Context) error {
-	if err := t.stopService(ctx, "boltz"); err != nil {
-		return err
+	if t.Network != types.Simnet {
+		if err := t.stopService(ctx, "boltz"); err != nil {
+			return err
+		}
 	}
 
 	if err := t.stopService(ctx, "xud"); err != nil {
 		return err
 	}
 
-	g, ctx := errgroup.WithContext(ctx)
+	g, _ := errgroup.WithContext(ctx)
 	g.Go(func() error {
 		return t.stopService(ctx, "lndbtc")
 	})
@@ -54,6 +85,7 @@ func (t *Launcher) down(ctx context.Context) error {
 		return err
 	}
 	defer os.Chdir(wd)
+
 	if err := os.Chdir(t.NetworkDir); err != nil {
 		return err
 	}
@@ -63,20 +95,31 @@ func (t *Launcher) down(ctx context.Context) error {
 			return err
 		}
 	}
-
 	c := exec.Command("docker-compose", "down")
 	c.Stdout = os.Stdout
 	c.Stderr = os.Stderr
-
 	return c.Run()
 }
 
 func (t *Launcher) removeFiles(ctx context.Context) error {
+	t.Logger.Debugf("homedir=%s", t.HomeDir)
 	if err := os.Chdir(t.HomeDir); err != nil {
 		return err
 	}
-	if err := os.RemoveAll(t.NetworkDir); err != nil {
+	fmt.Printf("Do you want to remove all %s files (%s)? [y/N] ", t.Network, t.NetworkDir)
+	var reply string
+	_, err := fmt.Scanln(&reply)
+	if err != nil {
+		if runtime.GOOS == "windows" && err.Error() == "unexpected newline" {
+			return nil
+		}
 		return err
+	}
+	reply = strings.ToLower(reply)
+	if reply == "y" || reply == "yes" {
+		if err := os.RemoveAll(t.NetworkDir); err != nil {
+			return err
+		}
 	}
 	return nil
 }
